fix : getPost params cast

+++ inventory.php
@@ -83,8 +83,8 @@ function _action()
 
             $fk_inventory = $inventory->save($PDOdb);
             $fk_category = GETPOST('fk_category', 'array');
-            $fk_supplier = GETPOST('fk_supplier','int');
-            $fk_warehouse = GETPOST('fk_warehouse','int');
+            $fk_supplier = (int)GETPOST('fk_supplier','int');
+            $fk_warehouse = (int)GETPOST('fk_warehouse','int');
 			$only_prods_in_stock = GETPOST('OnlyProdsInStock','int');
 			$inventoryWithBatchDetail = GETPOST('inventoryWithBatchDetail','int');
 			$inventory->per_batch = $inventoryWithBatchDetail;
